# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vendus REST API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Get list of suppliers",
                "description": "Returns paginated list of suppliers with navigation links.",
                "operationId": "getSuppliers",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter suppliers by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filter suppliers by exact email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "max_due_days",
                        "in": "query",
                        "description": "Filter suppliers by maximum due days",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "contract_file",
                        "in": "query",
                        "description": "Filter suppliers by contract file presence",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Supplier"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "description": "Link to the first page",
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "description": "Link to the last page",
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "description": "Link to the previous page",
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "description": "Link to the next page",
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "description": "Current page number",
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "description": "First item number on the current page",
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "description": "Last page number",
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "description": "Base path for the paginated links",
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "description": "Number of items per page",
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "description": "Last item number on the current page",
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "description": "Total items available",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create a new supplier",
                "description": "Creates a new supplier and returns the newly created supplier.",
                "operationId": "addSupplier",
                "requestBody": {
                    "description": "Supplier data to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "max_due_days"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the supplier",
                                        "type": "string",
                                        "example": "XPTO"
                                    },
                                    "email": {
                                        "description": "Email address of the supplier",
                                        "type": "string",
                                        "format": "email",
                                        "example": "xpto@email.com"
                                    },
                                    "vat": {
                                        "description": "VAT number of the supplier",
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "max_due_days": {
                                        "description": "Maximum number of due days allowed",
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "contract_file": {
                                        "description": "Path to the contract file associated with the supplier",
                                        "type": "string",
                                        "example": "/files/contracts/Supplier_Contract.pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "The name of the supplier",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "The email address of the supplier",
                                            "type": "string"
                                        },
                                        "max_due_days": {
                                            "description": "The maximum number of due days allowed for the supplier",
                                            "type": "integer"
                                        },
                                        "contract_file": {
                                            "description": "The file path for the contract file associated with the supplier",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error messsage"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "Object with all the errors by field"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Supplier": {
                "required": [
                    "id",
                    "name",
                    "vat",
                    "email",
                    "address",
                    "max_due_days",
                    "contract_file"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the supplier",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "The name of the supplier",
                        "type": "string"
                    },
                    "vat": {
                        "description": "The VAT of the supplier",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email address of the supplier",
                        "type": "string"
                    },
                    "address": {
                        "description": "The address of the supplier",
                        "type": "string"
                    },
                    "max_due_days": {
                        "description": "The maximum number of due days allowed for the supplier",
                        "type": "integer"
                    },
                    "contract_file": {
                        "description": "The file path for the contract file associated with the supplier",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}